Fixed more compile errors
